backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
pyserial==3.5
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
python-multipart==0.0.12
aiofiles==24.1.0
